+++ packages/core/src/core.ts
@@ -361,6 +361,11 @@ export async function getIDToken(aud?: string): Promise<string> {
 }
 
 /**
- * Markdown summary exports
+ * Summary exports
  */
-export {markdownSummary} from './markdown-summary'
+export {summary} from './summary'
+
+/**
+ * @deprecated use core.summary
+ */
+export {markdownSummary} from './summary'